requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
bootstrap4==0.1.0
confusable-homoglyphs==3.2.0
Django==3.0.8
django-bootstrap4==2.2.0
django-ckeditor==5.9.0
django-crispy-forms==1.9.2
django-js-asset==1.2.2
django-registration==3.1
django-widget-tweaks==1.4.8
Pillow==7.2.0
pytz==2020.1
soupsieve==2.0.1
sqlparse==0.3.1
